==> Optimization/notebooks/4_create_combined_df.ipynb <==
### Libraries

import pandas as pd

# FUNCTION: Combined DataFrame

### INPUT DATA: OPTIMIZED VEHICLES DFs 
### OUTPUT DATA: COMBINED VEHICLES DF

def combine_optimized_dfs(*dfs):
    return pd.concat(dfs, axis=1)

# Add DATA

max_space_vehicles = pd.read_csv('data/optimized_spatial_25_1503.csv') 
max_pop_vehicles = pd.read_csv('data/optimized_pop_count_25_1503.csv')
fair_vehicles = pd.read_csv('data/optimized_fairness_25_1503.csv') 

# Usage Function
combined_df = combine_optimized_dfs(max_space_vehicles, max_pop_vehicles, fair_vehicles)

#combined_df.to_csv('data/combined_vehicles_1503.csv', index=False)
